+++ integration/spark/app/src/test/resources/spark_scripts/spark_overwrite_hive.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
+import time
 
 from pyspark.sql import SparkSession
 
@@ -21,3 +22,5 @@ spark.sql("CREATE TABLE IF NOT EXISTS test (key INT, value STRING) USING hive")
 
 spark.sql("INSERT OVERWRITE DIRECTORY '/tmp/overwrite/table' USING hive VALUES (1, 'a'), (2, 'b'), (3, 'c')")
 result = spark.sql("SELECT count(*) from test")
+
+time.sleep(1)